Move enum Color, PieceName into Piece class rename enum PieceName to Type
