--- src/main/java/chess/pieces/Color.java
@@ -1,6 +1,0 @@
-package chess.pieces;
-
-public enum Color {
-    WHITE,
-    BLACK
-}
